requirements.txt
argparse
cryptography
#ed25519
pynacl